separate logic for query and event listeners
